test/requirements.txt
Appium-Python-Client==0.47
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
decorator==4.4.0
docutils==0.15.2
flex==6.14.1
genson==1.1.0
idna==2.8
importlib-metadata==0.23
jsonpath-ng==1.4.3
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonpointer==2.0
jsonschema==3.0.2
kitchen==1.2.6
more-itertools==7.2.0
packaging==19.1
pbr==5.4.3
pluggy==0.13.0
ply==3.11
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
RESTinstance==1.0.2
rfc3987==1.3.8
robotframework==3.1.2
robotframework-requests==0.5.0
robotframework-seleniumlibrary==3.3.1
selenium==3.141.0
six==1.12.0
strict-rfc3339==0.7
tox==3.0.0
tzlocal==2.0.0
urllib3==1.25.3
validate-email==1.3
virtualenv==16.7.5
wcwidth==0.1.7
zipp==0.6.0
browserstack-local==1.2.2
python-dotenv==0.12.0